# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/categories": {
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new category",
        "tags": [
          "Categories"
        ]
      },
      "get": {
        "operationId": "CategoryController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for filtering",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "required": false,
            "in": "query",
            "description": "Filter categories by title (case-insensitive partial match)",
            "schema": {
              "example": "work",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of categories.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a list of categories for the current user",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The category details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve a single category by ID",
        "tags": [
          "Categories"
        ]
      },
      "patch": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The category has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing category",
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The category has been successfully soft-deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete a category by ID",
        "tags": [
          "Categories"
        ]
      }
    }
  },
  "info": {
    "title": "AppStart API",
    "description": "API REST do AppStart - Especificação OpenAPI 3.0",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "appstart_session",
        "description": "Cookie de sessão local opaca"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 100
          },
          "totalPages": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the category",
            "example": "Work",
            "minLength": 3,
            "maxLength": 50
          }
        },
        "required": [
          "title"
        ]
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the category",
            "example": "clx0x0x0x0x0x0x0x0x0x0x0"
          },
          "title": {
            "type": "string",
            "description": "The title of the category",
            "example": "Work"
          },
          "ownerId": {
            "type": "string",
            "description": "The ID of the user who owns this category",
            "example": "clx0x0x0x0x0x0x0x0x0x0x0"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the category was created",
            "example": "2023-10-27T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the category was last updated",
            "example": "2023-10-27T11:00:00.000Z"
          },
          "deletedAt": {
            "type": "object",
            "description": "Timestamp when the category was soft-deleted",
            "example": "2023-10-27T12:00:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "ownerId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array of items",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "description": "Pagination metadata",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetaDto"
              }
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the category",
            "example": "Work",
            "minLength": 3,
            "maxLength": 50
          }
        }
      }
    }
  }
}